sqrt -> sqrtf

+++ common/vec3.h
@@ -57,7 +57,7 @@ public:
     return Vec3(x * o.x, y * o.y, z * o.z);
   }
   float len2() const { return x*x + y*y + z*z; }
-  float len() const { return sqrt(len2()); }
+  float len() const { return sqrtf(len2()); }
 
   static void Rotate90(float& a, float& b) {
     float tmp = b;
